Do not close a reconnected websocket
If you click on a download link in Firefox, then the websocket is disconnected briefly until Firefox realizes it is a download request. The reconnect then happens in less than 1 second so the new connection will be closed on the server after the timeout of 1s has elapsed (when the code tries to close an old connection). After this the websocket connection will not work.